classification/requirements.txt
fastapi==0.109.1
uvicorn==0.17.6
mysql-connector-python==8.0.23
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
pywavelets==1.6.0
scikit-learn==1.5.0
kneed == 0.8.5
lightgbm